+++ src/App/Commands/Databases/DatabasesDescribeCommand.php
@@ -1,7 +1,7 @@
 <?php
 
 
-namespace Console\App\Commands\Database;
+namespace Console\App\Commands\Databases;
 
 
 use Art4\JsonApiClient\Helper\Parser;
@@ -10,22 +10,17 @@ use Art4\JsonApiClient\V1\Document;
 use Console\App\Commands\Command;
 use Exception;
 use GuzzleHttp\Exception\GuzzleException;
-use InvalidArgumentException;
 use Symfony\Component\Console\Helper\Table;
-use Symfony\Component\Console\Helper\TableSeparator;
+use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
-use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
+use InvalidArgumentException;
 
-class DatabaseListCommand extends Command
+class DatabasesDescribeCommand extends Command
 {
-	protected static $defaultName = 'database:list';
+	protected static $defaultName = 'databases:describe';
 
-	const API_ENDPOINT = 'https://api.lamp.io/databases%s';
-
-	const OPTIONS_TO_QUERY_KEYS = [
-		'organization_id' => 'filter[organization_id]',
-	];
+	const API_ENDPOINT = 'https://api.lamp.io/databases/%s';
 
 	/**
 	 *
@@ -33,9 +28,9 @@ class DatabaseListCommand extends Command
 	protected function configure()
 	{
 		parent::configure();
-		$this->setDescription('Returns all allowed databases')
-			->setHelp('https://www.lamp.io/api#/databases/databasesList')
-			->addOption('organization_id', null, InputOption::VALUE_REQUIRED, 'Filter output by organization id value');
+		$this->setDescription('Returns a database')
+			->setHelp('https://www.lamp.io/api#/databases/databasesShow')
+			->addArgument('database_id', InputArgument::REQUIRED, 'The id of database');
 	}
 
 	/**
@@ -50,8 +45,9 @@ class DatabaseListCommand extends Command
 		try {
 			$response = $this->httpHelper->getClient()->request(
 				'GET',
-				sprintf(self::API_ENDPOINT,
-					$this->httpHelper->optionsToQuery($input->getOptions(), self::OPTIONS_TO_QUERY_KEYS)
+				sprintf(
+					self::API_ENDPOINT,
+					$input->getArgument('database_id')
 				),
 				[
 					'headers' => $this->httpHelper->getHeaders(),
@@ -84,28 +80,18 @@ class DatabaseListCommand extends Command
 	{
 		$table->setHeaderTitle('Database');
 		$serializer = new ArraySerializer(['recursive' => true]);
-		$serializedDocument = $serializer->serialize($document);
-		$header = ['Id', 'Attributes'];
-		$table->setHeaders($header);
-		foreach ($serializedDocument['data'] as $key => $data) {
-			$row = [$data['id']];
-			$attributeArray = [];
-			foreach ($data['attributes'] as $attributeKey => $attribute) {
-				if ($attributeKey != 'my_cnf') {
-					$attributeArray[] = $attributeKey . ' : ' . wordwrap($attribute, 50);
-				} else {
-					$attributeArray[] = $attributeKey . ' : ' . $attribute;
-				}
-			}
-			$row[] = (implode(PHP_EOL, $attributeArray));
-			$table->addRow($row);
-			if ($key != count($serializedDocument['data']) - 1) {
-				$table->addRow(new TableSeparator());
-			}
-
+		$headers = ['Id'];
+		$row = [$document->get('data.id')];
+		foreach ($serializer->serialize($document)['data']['attributes'] as $attributeKey => $attribute) {
+			array_push(
+				$row, ($attributeKey != 'my_cnf') ? $attribute : wordwrap($attribute, 30)
+			);
+			array_push($headers, $attributeKey);
 		}
-		$table->setColumnWidth(1, 200);
+		$table->setHeaders($headers);
+		$table->addRow($row);
 		return $table;
 
 	}
 }
+
