+++ src/App/Commands/Databases/DatabasesNewCommand.php
@@ -1,29 +1,26 @@
 <?php
 
+namespace Console\App\Commands\Databases;
 
-namespace Console\App\Commands\Database;
-
-
-use Art4\JsonApiClient\Helper\Parser;
-use Art4\JsonApiClient\Serializer\ArraySerializer;
-use Art4\JsonApiClient\V1\Document;
 use Console\App\Commands\Command;
 use GuzzleHttp\Exception\GuzzleException;
 use InvalidArgumentException;
-use Symfony\Component\Console\Helper\Table;
-use Symfony\Component\Console\Input\InputArgument;
-use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
+use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
+use Art4\JsonApiClient\V1\Document;
+use Symfony\Component\Console\Helper\Table;
+use Art4\JsonApiClient\Serializer\ArraySerializer;
+use Art4\JsonApiClient\Helper\Parser;
 
-class DatabaseUpdateCommand extends Command
+class DatabasesNewCommand extends Command
 {
-	protected static $defaultName = 'database:update';
+	protected static $defaultName = 'databases:new';
 
-	const API_ENDPOINT = 'https://api.lamp.io/databases/%s';
+	const API_ENDPOINT = 'https://api.lamp.io/databases';
 
 	const EXCLUDE_FROM_OUTPUT = [
-		'my_cnf',
+		'my_cnf'
 	];
 
 	/**
@@ -32,16 +29,15 @@ class DatabaseUpdateCommand extends Command
 	protected function configure()
 	{
 		parent::configure();
-		$this->setDescription('Update a database')
-			->setHelp('https://www.lamp.io/api#/databases/databasesUpdate')
-			->addArgument('database_id', InputArgument::REQUIRED, 'The id of database')
-			->addOption('description', 'd', InputOption::VALUE_REQUIRED, 'Description of your database')
-			->addOption('memory', 'm', InputOption::VALUE_REQUIRED, 'Amount of virtual memory on your database')
-			->addOption('organization_id', null, InputOption::VALUE_REQUIRED, 'Name of your organization')
-			->addOption('my_cnf', null, InputOption::VALUE_REQUIRED, 'Path to your database config file')
-			->addOption('mysql_root_password', null, InputOption::VALUE_REQUIRED, 'Root password')
-			->addOption('ssd', null, InputOption::VALUE_REQUIRED, 'Size of ssd storage')
-			->addOption('vcpu', null, InputOption::VALUE_REQUIRED, 'The number of virtual cpu cores available');
+		$this->setDescription('Create a new database')
+			->setHelp('https://www.lamp.io/api#/databases/databasesCreate')
+			->addOption('description', 'd', InputOption::VALUE_REQUIRED, 'Description of your database', '')
+			->addOption('memory', 'm', InputOption::VALUE_REQUIRED, 'Amount of virtual memory on your database, default 512Mi', '512Mi')
+			->addOption('organization_id', null, InputOption::VALUE_REQUIRED, 'Name of your organization', '')
+			->addOption('my_cnf', null, InputOption::VALUE_REQUIRED, 'Path to your database config file', '')
+			->addOption('mysql_root_password', null, InputOption::VALUE_REQUIRED, 'Root password', '')
+			->addOption('ssd', null, InputOption::VALUE_REQUIRED, 'Size of ssd storage, default 1Gi', '1Gi')
+			->addOption('vcpu', null, InputOption::VALUE_REQUIRED, 'The number of virtual cpu cores available, default 0.25', '0.25');
 	}
 
 	/**
@@ -53,17 +49,10 @@ class DatabaseUpdateCommand extends Command
 	protected function execute(InputInterface $input, OutputInterface $output)
 	{
 		parent::execute($input, $output);
-
-		$this->getRequestBody($input);
-
 		try {
 			$response = $this->httpHelper->getClient()->request(
-				'PATCH',
-				sprintf(
-					self::API_ENDPOINT,
-					$input->getArgument('database_id')
-
-				),
+				'POST',
+				self::API_ENDPOINT,
 				[
 					'headers' => $this->httpHelper->getHeaders(),
 					'body'    => $this->getRequestBody($input),
@@ -97,7 +86,7 @@ class DatabaseUpdateCommand extends Command
 			throw  new InvalidArgumentException('Path to mysql config not valid');
 		}
 
-		$body = [
+		return json_encode([
 			'data' => [
 				'attributes' => array_merge([
 					'description'         => $input->getOption('description'),
@@ -107,18 +96,9 @@ class DatabaseUpdateCommand extends Command
 					'ssd'                 => $input->getOption('ssd'),
 					'vcpu'                => (float)$input->getOption('vcpu'),
 				], !empty($input->getOption('organization_id')) ? ['organization_id' => $input->getOption('organization_id')] : []),
-				'id'         => $input->getArgument('database_id'),
 				'type'       => 'databases',
 			],
-		];
-
-		$body['data']['attributes'] = array_filter($body['data']['attributes'], function ($val, $key) {
-			if (!empty($val)) {
-				return [$key => $val];
-			}
-		}, ARRAY_FILTER_USE_BOTH);
-
-		return json_encode($body);
+		]);
 	}
